first day comit
